Fix singleton tmp files cleanup
Only in singleton mode, directory cleaning needs to be done
by the program itself.
There are some problems with these parts of the code that cause
the directory to not be cleaned.
This commit fixes *some of* these issues.

1. btl/sm will not unlink its segments file. We never noticed this
in non-singleton mode because pmix cleaned it up for us.

After this, we can clean up the segment file created by sm in
/dev/shm.(when singletons normally terminated)

2. Modified the singleton session directory structure and enabled
recursive deletion.

After this, we can cleanup the session dir. (when singletons normally
terminated)

3. Fix a bug - local peer number of a singleton should be 0, not 1.

After this, the btl/sm and btl/smcuda components will return NULL
during their init process and will be automatically closed.
btl/sm segment file in /dev/shm will never be created in singleton mode
now.

4. If the singleton does a comm_spawn, register the singleton's session
directory for cleanup by the "prte".

Signed-off-by: xbw <78337767+xbw22109@users.noreply.github.com>